// src/graphql-schema.ts
export type ID = string
export type GitObjectID = string
// ...
export interface IPageInfo {
    __typename: 'PageInfo'
    endCursor: string | null
    hasNextPage: boolean
}
// ...
export interface IRepository {
    __typename: 'Repository'
    id: ID
    name: string

    /**
     * @deprecated "Use name."
     */
    uri: string
    description: string
    language: string
    createdAt: DateTime
    updatedAt: DateTime | null
    commit: IGitCommit | null
    mirrorInfo: IMirrorRepositoryInfo
    externalRepository: IExternalRepository
    isFork: boolean
    isArchived: boolean
    isPrivate: boolean
    externalServices: IExternalServiceConnection

    /**
     * @deprecated "use Repository.mirrorInfo.cloneInProgress instead"
     */
    cloneInProgress: boolean
    textSearchIndex: IRepositoryTextSearchIndex | null
    url: string
    externalURLs: IExternalLink[]
    defaultBranch: IGitRef | null
    gitRefs: IGitRefConnection
    branches: IGitRefConnection
    tags: IGitRefConnection
    comparison: IRepositoryComparison
    contributors: IRepositoryContributorConnection
    viewerCanAdminister: boolean
    icon: string
    label: IMarkdown
    detail: IMarkdown
    matches: ISearchResultMatch[]
    lsifUploads: ILSIFUploadConnection
    lsifIndexes: ILSIFIndexConnection
    authorizedUsers: IUserConnection
    permissionsInfo: IPermissionsInfo | null
}
// ...
export interface IMarkdown {
    __typename: 'Markdown'
    text: string
    html: string
}

export interface ISearchResultMatch {
    __typename: 'SearchResultMatch'
    url: string
    body: IMarkdown
    highlights: IHighlight[]
}

export interface IHighlight {
    __typename: 'Highlight'
    line: number
    character: number
    length: number
}

export interface IGitCommit {
    __typename: 'GitCommit'
    id: ID
    repository: IRepository
    oid: GitObjectID
    abbreviatedOID: string
    author: ISignature
    committer: ISignature | null
    message: string
    subject: string
    body: string | null
    parents: IGitCommit[]
    url: string
    canonicalURL: string
    externalURLs: IExternalLink[]
    tree: IGitTree | null
    blob: IGitBlob | null
    file: File2 | null
    languages: string[]
    languageStatistics: ILanguageStatistics[]
    ancestors: IGitCommitConnection
    behindAhead: IBehindAheadCounts
    symbols: ISymbolConnection
}
// ...
export interface ISignature {
    __typename: 'Signature'
    person: IPerson
    date: string
}

export interface IPerson {
    __typename: 'Person'
    name: string
    email: string
    displayName: string
    avatarURL: string
    user: IUser | null
}

export interface IExternalLink {
    __typename: 'ExternalLink'
    url: string
    serviceType: string | null
}

export interface IGitTree {
    __typename: 'GitTree'
    path: string
    isRoot: boolean
    name: string
    isDirectory: boolean
    commit: IGitCommit
    repository: IRepository
    url: string
    canonicalURL: string
    externalURLs: IExternalLink[]
    rawZipArchiveURL: string
    submodule: ISubmodule | null
    directories: IGitTree[]
    files: IFile[]
    entries: TreeEntry[]
    symbols: ISymbolConnection
    isSingleChild: boolean
    lsif: TreeEntryLSIFData | null
}
// ...
export interface ISymbolConnection {
    __typename: 'SymbolConnection'
    nodes: ISymbol[]
    pageInfo: IPageInfo
}

export interface ISymbol {
    __typename: 'Symbol'
    name: string
    containerName: string | null
    kind: SymbolKind
    language: string
    location: ILocation
    url: string
    canonicalURL: string
    fileLocal: boolean
}
// ...
export interface IGitBlob {
    __typename: 'GitBlob'
    path: string
    name: string
    isDirectory: boolean
    content: string
    byteSize: number
    binary: boolean
    richHTML: string
    commit: IGitCommit
    repository: IRepository
    url: string
    canonicalURL: string
    externalURLs: IExternalLink[]
    blame: IHunk[]
    highlight: IHighlightedFile
    submodule: ISubmodule | null
    symbols: ISymbolConnection
    isSingleChild: boolean
    lsif: IGitBlobLSIFData | null
}
// ...
export type File2 = IGitBlob | IVirtualFile
// ...
export interface IFile {
    __typename: 'File'
    path: string
    name: string
    isDirectory: boolean
    url: string
    repository: IRepository
}

export interface ILanguageStatistics {
    __typename: 'LanguageStatistics'
    name: string
    totalBytes: number
    totalLines: number
}

export interface IGitCommitConnection {
    __typename: 'GitCommitConnection'
    nodes: IGitCommit[]
    totalCount: number | null
    pageInfo: IPageInfo
}

export interface IBehindAheadCounts {
    __typename: 'BehindAheadCounts'
    behind: number
    ahead: number
}
// ...
export interface IGitRef {
    __typename: 'GitRef'
    id: ID
    name: string
    abbrevName: string
    displayName: string
    prefix: string
    type: GitRefType
    target: IGitObject
    repository: IRepository
    url: string
}

export enum GitRefType {
    GIT_BRANCH = 'GIT_BRANCH',
    GIT_TAG = 'GIT_TAG',
    GIT_REF_OTHER = 'GIT_REF_OTHER',
}

export interface IGitObject {
    __typename: 'GitObject'
    oid: GitObjectID
    abbreviatedOID: string
    commit: IGitCommit | null
    type: GitObjectType
}
// ...
export interface ISearch {
    __typename: 'Search'
    results: ISearchResults
    suggestions: SearchSuggestion[]
    stats: ISearchResultsStats
}

export interface ISuggestionsOnSearchArguments {
    first?: number | null
}

export interface ISearchResults {
    __typename: 'SearchResults'
    results: SearchResult[]
    matchCount: number

    /**
     * @deprecated "renamed to matchCount for less ambiguity"
     */
    resultCount: number
    approximateResultCount: string
    limitHit: boolean
    sparkline: number[]
    repositories: IRepository[]
    repositoriesCount: number
    repositoriesSearched: IRepository[]
    indexedRepositoriesSearched: IRepository[]
    cloning: IRepository[]
    missing: IRepository[]
    timedout: IRepository[]
    indexUnavailable: boolean
    alert: ISearchAlert | null
    elapsedMilliseconds: number
    dynamicFilters: ISearchFilter[]
    pageInfo: IPageInfo
}

export type SearchResult = IFileMatch | ICommitSearchResult | IRepository | ICodemodResult
// ...
export interface ICommitSearchResult {
    __typename: 'CommitSearchResult'
    icon: string
    label: IMarkdown
    url: string
    detail: IMarkdown
    matches: ISearchResultMatch[]
    commit: IGitCommit
    refs: IGitRef[]
    sourceRefs: IGitRef[]
    messagePreview: IHighlightedString | null
    diffPreview: IHighlightedString | null
}

export interface IHighlightedString {
    __typename: 'HighlightedString'
    value: string
    highlights: IHighlight[]
}

export interface ICodemodResult {
    __typename: 'CodemodResult'
    icon: string
    label: IMarkdown
    url: string
    detail: IMarkdown
    matches: ISearchResultMatch[]
    commit: IGitCommit
    rawDiff: string
}

export interface ISearchAlert {
    __typename: 'SearchAlert'
    title: string
    description: string | null
    proposedQueries: ISearchQueryDescription[] | null
}

export interface ISearchQueryDescription {
    __typename: 'SearchQueryDescription'
    description: string | null
    query: string
}

export interface ISearchFilter {
    __typename: 'SearchFilter'
    value: string
    label: string
    count: number
    limitHit: boolean
    kind: string
}

export type SearchSuggestion = IRepository | IFile | ISymbol | ILanguage

export interface ILanguage {
    __typename: 'Language'
    name: string
}

export interface ISearchResultsStats {
    __typename: 'SearchResultsStats'
    approximateResultCount: string
    sparkline: number[]
    languages: ILanguageStatistics[]
}